requirements.txt
fastapi
uvicorn[standard]
python-socketio[client]
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-genai
pydantic-settings
